In verbatim '<<<', ignore regular markup '<tag|...>'
But nested <<<tag|...>>> can be found in there.

Reminder: verbatim markup doesn't mean there is no parsing inside at
all: it means the parsing is restricted to triple '<<<' markup.

The content inside a verbatim tag can be freely interpreted in a
verbatim way (like <pre></pre>) or not by the user of nxml.

It's only a convenience notation to avoid escaping.