nxml: stricter whitespace handling
The rules on whitespaces are:

Always strip leading indentation whitespaces, whether in verbatim mode
or not.

All leading whitespaces are ignored in blocks started with a newline
opener (called "implicit opener").

All trailing whitespaces are ignored in blocks ended with a newline
closer (called "implicit closer").

Preserve leading whitespaces in blocks with a same-line explicit opener
"<tag|".

Preserve trailing whitespaces in blocks with an explicit closer.

With these rules, it is possible to generate XML with precise control of
where white spaces appear. With HTML output, this is important, as all
sorts of corner cases exist where whitespaces are actually relevant
(e.g. a newline before a </pre>).

The default behavior with implicit openers and closers does the "right"
thing to generate HTML.